requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2

# Machine Learning & Model Serving
# Note: tensorflow/keras installation may require special setup
# For CPU-only: pip install tensorflow
# For GPU: pip install tensorflow[and-cuda]
# Alternative: pip install keras
keras>=3.0.0

# Image Processing
pillow==10.0.0
numpy>=1.20.0,<1.25.0

# Data Processing
pandas>=1.0.0

# Testing
pytest==7.4.2
httpx==0.25.0

# Development & Deployment
h5py>=3.0.0

# Optional: For production deployments
# gunicorn==21.2.0  # Production ASGI server
# python-dotenv==1.0.0  # Environment variable management
